refactored to decouple withdraw and check from main
